safari-534.53-branch/LayoutTests/http/tests/loading/text-content-type-with-binary-extension-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
frame "iframe" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
frame "iframe" - didFailProvisionalLoadWithError
main frame - didFinishLoadForFrame
Test for https://bugs.webkit.org/show_bug.cgi?id=17360 REGRESSION: mp4 file downloaded from server is downloaded as html.

Binary data not loaded as text. Maybe PASS.
